EmlakCRM/backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0
redis==5.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# AWS S3
boto3==1.35.36

# Web Scraping (bina.az, tap.az parser)
httpx==0.27.2
beautifulsoup4==4.12.3
selenium==4.26.1

# WhatsApp Integration
twilio==9.3.7

# Image Processing
Pillow==11.0.0

# Utilities
python-dotenv==1.0.1
